Modify and finished logic for login & register services

+++ server/server.js
@@ -16,7 +16,7 @@ app.get('/api/users', (req, res) => {
   res.json(JSON.parse(data));
 });
 
-app.post('/api/users', (req, res) => {
+app.post('/api/users/register', (req, res) => {
   const users = JSON.parse(fs.readFileSync(DATA_FILE));
   const newUser = {
     id: req.body.id,
@@ -24,12 +24,43 @@ app.post('/api/users', (req, res) => {
     email: req.body.email,
     password: req.body.password,
   };
-  users.push(newUser);
-  fs.writeFileSync(DATA_FILE, JSON.stringify(users));
 
-  const userResponse = { ...newUser };
+  const alreadyRegisteredUser = users.some((user) => user.email === newUser.email);
+
+  if (!alreadyRegisteredUser) {
+    users.push(newUser);
+    fs.writeFileSync(DATA_FILE, JSON.stringify(users));
+
+    const userResponse = { ...newUser };
+    delete userResponse.password;
+
+    res.status(201).json({ user: userResponse });
+  } else {
+    res.status(401);
+    throw new Error('User already exists');
+  }
+});
+
+app.post('/api/users/login', (req, res) => {
+  const users = JSON.parse(fs.readFileSync(DATA_FILE));
+  const requestedUser = {
+    email: req.body.email,
+    password: req.body.password,
+  };
+
+  const loggedUser = users.find(
+    (user) => user.email === requestedUser.email && user.password === requestedUser.password
+  );
+
+  const userResponse = { ...loggedUser };
   delete userResponse.password;
-  res.status(201).json({ user: userResponse });
+
+  if (loggedUser.length > 0) {
+    res.status(201).json({ user: userResponse });
+  } else {
+    res.status(401);
+    throw new Error('Invalid email or password');
+  }
 });
 
 const notFound = (req, res, next) => {
